+++ mslicer/src/app/config/sliced.rs
@@ -7,6 +7,10 @@ pub struct SlicedConfig {
     pub view: SlicePreviewView,
     pub multisample: u32,
     pub sidebar: bool,
+
+    pub resin_density: f32, // g/mL
+    pub resin_cost: f32,    // ¤/mL
+    pub currency: Currency,
 }
 
 #[derive(Default, PartialEq, Eq, Copy, Clone, Serialize, Deserialize)]
@@ -21,6 +25,15 @@ pub enum SlicePreviewView {
     Screen,
     #[default]
     BuildPlate,
+}
+
+#[derive(Default, PartialEq, Eq, Copy, Clone, Serialize, Deserialize)]
+pub enum Currency {
+    Euro,
+    #[default]
+    Dollar,
+    Yen,
+    Generic,
 }
 
 impl SlicePreviewCoordinateSpace {
@@ -45,13 +58,39 @@ impl SlicePreviewView {
     }
 }
 
+impl Currency {
+    pub const ALL: [Self; 4] = [Self::Euro, Self::Dollar, Self::Yen, Self::Generic];
+
+    pub fn name(&self) -> &str {
+        match self {
+            Currency::Euro => "Euro",
+            Currency::Dollar => "Dollar",
+            Currency::Yen => "Yen",
+            Currency::Generic => "Generic",
+        }
+    }
+
+    pub fn symbol(&self) -> char {
+        match self {
+            Currency::Euro => '€',
+            Currency::Dollar => '$',
+            Currency::Yen => '¥',
+            Currency::Generic => '¤',
+        }
+    }
+}
+
 impl Default for SlicedConfig {
     fn default() -> Self {
         Self {
-            coordinate_space: SlicePreviewCoordinateSpace::WorldSpace,
-            view: SlicePreviewView::BuildPlate,
+            coordinate_space: Default::default(),
+            view: Default::default(),
             multisample: 8,
             sidebar: true,
+
+            resin_density: 1.1,
+            resin_cost: 19.79,
+            currency: Currency::default(),
         }
     }
 }